+++ src/main/java/model/participant/Participants.java
@@ -1,11 +1,8 @@
-package model;
+package model.participant;
 
 import java.util.ArrayList;
 import java.util.Iterator;
 import java.util.List;
-import model.participant.Dealer;
-import model.participant.Participant;
-import model.participant.Player;
 
 public class Participants implements Iterable<Participant> {
     private final List<Participant> values;
